Wrote a separate, simple system for handling block-floor and block-block collisions
Spent a lot of time trying to get stacks of blocks balancing one on top
of the other strictly as a result of a general purpose collision system
(based on penetrating bounding boxes, and which would also work for,
e.g., box-bullet collisions) -- and finally realized that this is a
difficult problem that will consume all my time if I continue pursuing
it.

Instead I wrote a separate system for handling box stacking, which just
uses the number of boxes in a given column and a box height.